Add: Rename project, refactor code and reach quite stable state
A bunch of features have been added and the code have been almost
entirely refactored.

Here is a non exhaustive list of new features:

- Error code available
- Prompt color changes based on error code
- Log feature is now functional
- Some basic (and still buggy) completion, filtering with skim
- Hints based on history (like fish or zsh-auto-suggestions)

There is still a lot of work to do regarding splitting code across
several files, but it will be performed in later commits.

The project has been renamed to OSH since it has now reached a state
where it could be used as a shell for almost all every day cases.